extracted the date and time function from the tile and placed it inside common/event_time.dart for readability.
