+++ lib/ui/events/event_tile.dart
@@ -5,9 +5,10 @@ import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:intl/intl.dart';
 
+import '../common/event_time.dart';
+
 class EventTile extends StatelessWidget {
   const EventTile({Key? key, required this.data}) : super(key: key);
-
   /// STATES
   final double tileWidth = 190;
 
@@ -25,7 +26,7 @@ class EventTile extends StatelessWidget {
   Widget buildEventTile(BuildContext context) {
     return Container(
       width: tileWidth,
-      height: 300,
+      height: 260,
       margin: EdgeInsets.zero,
       child: InkWell(
         onTap: () {
@@ -63,7 +64,7 @@ class EventTile extends StatelessWidget {
                         Padding(
                           padding: EdgeInsets.only(bottom: 5),
                         ),
-                        eventsDateTime(data),
+                        EventTileDateTime(data: data),
                       ],
                     ),
                   ),
@@ -106,88 +107,4 @@ class EventTile extends StatelessWidget {
           );
   }
 
-  Widget eventsDateTime(EventModel data) {
-    try {
-      // Separate dates from times
-      var startMonthDayYear = DateFormat.yMMMMd('en_US').format(data.startDate.toLocal());
-      var endMonthDayYear = DateFormat.yMMMMd('en_US').format(data.endDate.toLocal());
-      var startTime = DateFormat.jm().format(data.startDate.toLocal());
-      var endTime = DateFormat.jm().format(data.endDate.toLocal());
-
-
-    // Mark any special types of events
-      var sameDay = (startMonthDayYear == endMonthDayYear);
-      var unspecifiedTime = (startTime == '12:00 AM' && endTime == '12:00 AM');
-      Widget date, time;
-      if (sameDay) {
-        date = Text(
-          startMonthDayYear,
-          style: TextStyle(fontSize: 12),
-        ); // Ex. June 11, 2021
-      } else {
-        // if not the same date, check if the same year
-        var startYear = startMonthDayYear.substring(
-            startMonthDayYear.indexOf(',') + 2, startMonthDayYear.length);
-        var endYear = endMonthDayYear.substring(
-            endMonthDayYear.indexOf(',') + 2, endMonthDayYear.length);
-        if (startYear == endYear) {
-          // if the same year, check if the same month
-          var startMonth = startMonthDayYear.substring(0, startMonthDayYear.indexOf(' '));
-          var endMonth = endMonthDayYear.substring(0, endMonthDayYear.indexOf(' '));
-          if (startMonth == endMonth) {
-            // if different date in the same month and year
-            var startDay = startMonthDayYear.substring(
-                startMonthDayYear.indexOf(' ') + 1,
-                startMonthDayYear.indexOf(','));
-            var endDay = endMonthDayYear.substring(
-                endMonthDayYear.indexOf(' ') + 1, endMonthDayYear.indexOf(','));
-            date = Text(
-              startMonth + ' ' + startDay + ' - ' + endDay + ', ' + startYear,
-              style: TextStyle(fontSize: 12),
-            ); // Ex. September 11 - 26, 2021
-          } else {
-            // if different month in the same year
-            var startMonthDay = startMonthDayYear.substring(0, startMonthDayYear.indexOf(','));
-            var endMonthDay = endMonthDayYear.substring(0, endMonthDayYear.indexOf(','));
-            date = Text(
-              startMonthDay + ' - ' + endMonthDay + ', ' + startYear,
-              style: TextStyle(fontSize: 12),
-            ); // Ex. September 11 - October 26, 2021
-          }
-        } else {
-          date = Text(
-            startMonthDayYear + ' - ' + endMonthDayYear,
-            style: TextStyle(fontSize: 12),
-          ); // Ex. June 11, 2021 - May 12, 2023
-        }
-      }
-
-      if (unspecifiedTime) {
-        time = Text(
-          '',
-          style: TextStyle(fontSize: 12),
-        );
-      } else {
-        time = Text(
-          startTime + ' - ' + endTime,
-          style: TextStyle(fontSize: 12),
-        );
-      }
-
-      return Column(
-        // mainAxisAlignment: MainAxisAlignment.center,
-        crossAxisAlignment: CrossAxisAlignment.center,
-        children: [
-          date,
-          Padding(
-            padding: EdgeInsets.only(bottom: 5),
-          ),
-          time
-        ],
-      );
-    } catch (e) {
-      print(e);
-      return Container();
-    }
-  }
 }
