+++ lib/ui/events/events_list.dart
@@ -53,10 +53,5 @@ class EventsList extends StatelessWidget {
             : EventsAll(),
       );
     }
-
-    // ListView(
-    //   children:
-    //   ListTile.divideTiles(tiles: eventTiles, context: context)
-    //       .toList(),
   }
 }
